+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- BlackboardContainer is renamed to SimpleBlackboardContainer.
+- SerializedContainer is renamed to SimpleSerializedContainer.
+- SerializedTablesContainer is renamed to SimpleSerializedTablesContainer.
+
 ## [1.3.0] - 2021-01-23
 
 ### Added